requirements.txt
Django==5.0
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
celery==5.3.6
django-celery-results==2.5.1
django-environ==0.11.2
gunicorn==22.0.0
rest_framework_nested==0.93.4
